import initial PoC version
Import the initial proof of concept code. At the moment, only TCP_STREAM
and TCP_RR tests are available, output is very raw, collecting server stats
is missing, error handling needs a lot of improvement and a lot of other
features are missing. But it is at least able to run simple tests.