client: more reliable sent/received data accounting
Using pthread_cancel() to kill client worker threads seemed convenient but
it has one sever drawback: when the thread is cancelled in send() syscall,
it is cancelled immediately even if the syscall may have already sent part
of the data which was or will be received by server but it is not accounted
for in client side stats. This is almost always the case in TCP_STREAM test
with default 1MB message size, resulting in raw data showing server Rx byte
count higher than client Tx which is plain wrong (server receiving more
data than client sent).

Switch to more tedious but cleaner way of signalling worker threads with
USR1 which is ignored but interrupts send() and recv() calls so that we can
count partial last send/receive in client stats. As a bonus, we can drop
the final wait for WS_FINISHED state (never triggered) which was meant to
keep the worker alive - but it was skipped in case of error anyway so that
it did not actually work that way.