server: do not send bogus reply after receive eof
When server recv_msg() encounters "eof" (i.e. connection closed by client),
it does not return an error but only sets the eof flag. When this happens,
we must not send a reply as we did not actually read a request message so
that there is nothing to reply to.

This bug resulted in "server send" numbers in TCP_RR raw data output being
one message higher than other three.