client: support for human readable results
Support for printing all results except raw data in human readable form,
i.e. in suitable multiples of B/s or tr/s. The unit is chosen to be the
biggest for which the displayed value is at least 20.0.

Exact results and both decadic (10^3) and binary (2^10) prefixes are
supported. (Corresponding command line options to be added next.)