client: calculate and display confidence interval after each iteration
After each iteration, calculate the confidence interval and print it as
part of the iteration summary. For now, 95% level interval is shown but
everything is prepared also for 99% confidence level.

A table of Student distribution p values is used to calculate the
confidence interval width up to 101 iterations. For higher number of
iterations (which should be rare), an asymptotic approximation is used
which should provide results with precision comparable to those in the
table.