silence compiler warnings
These only show when optimization is enabled so they were missed earlier.
The possibly uninitialized variables are in fact false positives but it may
not be obvious to the compiler and neither is a time critical code so there
is no harm adding an explicit initialization.

The missing field initializer in setup_listener() is also bogus as this
warning is not supposed to show when named struct initializers are used,
probably gcc got confused byt combination of htons() and function argument
in the initializer.